Target FPS
Crudely attempt to match target FPS by sleeping excess time away between
loops. (This does not perfectly track it and is not designed to, it is
more designed to lighten the cpu load by decreasing the amount of
frames drawn.) This implmenetation does have the side effect that lower
fps counts cause lag, this is not a huge problem if the mouse is not
used.